Problem: some compilers fail to compile sqlite
```
error: use of deleted function ‘static cppgres::datum cppgres::datum_conversion<T, <template-parameter-1-2> >::into_datum(const T&)
[with T = long long int; <template-parameter-1-2> = void]’
```

Solution: explicitly cast into int64_t

+++ extensions/omni_sqlite/omni_sqlite.cpp
@@ -168,7 +168,8 @@ std::generator<cppgres::record> query_results(int sqlite_rc, int column_count, s
         auto value_type = sqlite3_value_type(value);
         switch (value_type) {
         case SQLITE_INTEGER:
-          datums.emplace_back(cppgres::into_nullable_datum(sqlite3_column_int64(stmt, i)));
+          datums.emplace_back(
+              cppgres::into_nullable_datum(static_cast<int64_t>(sqlite3_column_int64(stmt, i))));
           break;
         case SQLITE_FLOAT:
           datums.emplace_back(cppgres::into_nullable_datum(sqlite3_column_double(stmt, i)));
